My dlmas page, Dashboard design, question description
Created a page (view, controller, route) that display all the created dlma by the user.
Improved the design of the dashboard. The dashboard now only display button to redirect user on all the functionalities of the site.
If the description of the question doesn't exists a message is displayed for the user to understand it.